detective step audio and room improvements
