+++ Assets/Materials/Black.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8587bc2e156521d43825e5a795ae846c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
