+++ Assets/Scripts/Detective.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edf554d46b10e6744b14d0247f53a0ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
