+++ Assets/Scripts/PlayerController.cs
@@ -57,7 +57,7 @@ public class PlayerController : MonoBehaviour
     // If this is P1, make them the Detective
     if (playerId == 0)
     {
-      GameObject.Find("Detective").GetComponent<PlayerMovement>().assignInputHandler(_inputHandler);
+      GameObject.Find("Detective").GetComponent<Detective>().assignInputHandler(_inputHandler);
       playerInput.SwitchCurrentActionMap("Detective");
     }
     else
