+++ ProjectSettings/EditorBuildSettings.asset
@@ -8,16 +8,13 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Menu.unity
     guid: 8f052ad4e09f76d44a3006f831a7469e
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/End.unity
     guid: 34b10751cfaa4a840baa0dbd96eb500e
   - enabled: 1
     path: Assets/Scenes/_MAINSCENE.unity
     guid: 74f598e239b39f04fb0e3ab5fc502fce
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
+  - enabled: 1
+    path: Assets/Scenes/_pre-playtest.unity
+    guid: 5dd52e0fa01ecdf43b7a7662088ab103
   m_configObjects: {}
